Rename tokensInFunctions to fixAlphaModificationFunctions
