+++ packages/stylelint-plugin/src/rules/lib/fixAlphaModificationFunctions.ts
